+++ src/lib/getContainer.ts
@@ -63,6 +63,8 @@ import {MoCMeasurementDataRepository} from '../repositories/fetchers/MoCMeasurem
 import {MoCMeasurementFetcher} from '../workers/fetchers/MoCMeasurementFetcher.js';
 import {BobDeviationDispatcher} from '../services/dispatchers/networks/BobDeviationDispatcher.js';
 import DataPurger from '../services/DataPurger.js';
+import {BinanceCandlestickRepository} from '../repositories/fetchers/BinanceCandlestickRepository.js';
+import {BinanceCandlestickFetcher} from '../workers/fetchers/BinanceCandlestickFetcher.js';
 
 export function getContainer(): Container {
   const container = new Container({autoBindInjectable: true});
@@ -82,6 +84,8 @@ export function getContainer(): Container {
 
   container.bind(BinancePriceGetter).toSelf().inSingletonScope();
   container.bind(BinanceDataRepository).toSelf().inSingletonScope();
+  container.bind(BinanceCandlestickRepository).toSelf().inSingletonScope();
+  container.bind(BinanceCandlestickFetcher).toSelf().inSingletonScope();
   container.bind(ByBitPriceGetter).toSelf().inSingletonScope();
   container.bind(ByBitDataRepository).toSelf().inSingletonScope();
   container.bind(CoingeckoDataRepository).toSelf().inSingletonScope();